android: add profile module + rework editor
